src/main/test/java/com/aoc/y2022/ProblemDay15Test.java
package com.aoc.y2022;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay15Test implements ProblemDayTest {

    private ProblemDay15 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay15();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(5299855);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(13615843289729L);
    }

}
